impl

+++ src/main/kotlin/io/github/t45k/scm/entity/TokenInfo.kt
@@ -1,5 +1,5 @@
 package io.github.t45k.scm.entity
 
-data class TokenInfo(val tokenNumber: Int, val lineNumber: Int)
+data class TokenInfo(val tokenNumber: Long, val lineNumber: Int)
 
-infix fun Int.to(lineNumber: Int): TokenInfo = TokenInfo(this, lineNumber)
+infix fun Long.to(lineNumber: Int): TokenInfo = TokenInfo(this, lineNumber)
